+++ app/Config/Routes.php
@@ -68,6 +68,7 @@ $routes->group('admin', ['filter' => 'AuthCheck'], function($routes){
 		$routes->add('store', 'Admin\Blog::store');
 		$routes->add('imagem', 'Admin\Blog::imagem');
 		$routes->add('upload', 'Admin\Blog::upload');
+		$routes->add('about', 'Admin\Blog::about');
 	});
 
 });